harness: rename testing package to "harness"
